# src/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/api/articles/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get Article List",
                "parameters": [
                    {},
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":[{\"id\":1,\"title\":\"abc\"}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get article",
                "parameters": [
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"id\":1,\"name\":\"\"}]",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Add User",
                "parameters": [
                    {},
                    {},
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":1}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/delete": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete User",
                "parameters": [
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200, \"msg\": \"操作成功\", \"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get User List",
                "parameters": [
                    {},
                    {}
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{\"id\":1,\"name\":\"Mantis\", \"email\":\"tangchunlinit@gmail.com\"}]",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}